Accesibility Issue Fixed
Found that on smaller screens we had text flowing out of their designated container when putting a large amount of text. Updated this to add scroll bards on those text sections. Also added scroll bars to the site in general so it is easier for new users to use. Still have to fix some marginal issue but I believe it is a caching issue. Will attempt to fix later