Fixed those margin issues
One was just a naming issue between css files. Believe it is a caching issue but a name fix fixed the whole thing. Then for the about us section I just used a div to fix up the button in the about us section. Think there is a better way to do it but for now this is a band-aid