Untested Js to interact with database
This isn't fully complete and needs to be linked to the home.js page or whatever page is being used to pull the ctf data. This has limited security and I'll try to fix it to make it act more like an API but for now this should work using my db access and once linked should be able to post, get, put, and delete the JSON data.